+++ NEWS.md
@@ -1,5 +1,9 @@
 # RPresto 1.4.1.9000
 
+* Add a convenient wrapper `kerberos_configs()` to generate Kerberos header
+  configs that can be passed to the `request.config` argument of `dbConnect()`.
+  (#202 and #221). Thanks to @suzzettedetorres for providing this solution.
+
 # RPresto 1.4.1
 
 * `dbListFields` now works with identifier name which in turn supports the use