+++ NEWS.md
@@ -1,4 +1,4 @@
-# RPresto 1.4.1.9000
+# RPresto 1.4.2
 
 * Add a convenient wrapper `kerberos_configs()` to generate Kerberos header
   configs that can be passed to the `request.config` argument of `dbConnect()`.